fixed audiofix autooff/on

+++ beta/App/Gmu/launch.sh
@@ -1,16 +1,30 @@
 #!/bin/sh
 
 my_dir=`dirname $0`
+runsvr=`/customer/app/jsonval audiofix`
 
-killall audioserver
+if [ "$runsvr" != "0" ] ; then
+    FILE=/customer/app/axp_test
+
+    if [ -f "$FILE" ]; then
+		printf "audioserver desactivado en MMP"
+        killall audioserver
+		killall audioserver.plu
+    else
+		printf "audioserver desactivado en MM"
+        killall audioserver
+		killall audioserver.min
+    fi
+fi
 
 cd $my_dir
 ./launch2.sh
 
 runsvr=`/customer/app/jsonval audiofix`
 if [ "$runsvr" != "0" ] ; then
-    /mnt/SDCARD/Koriki/bin/audioserver > /dev/null 2>&1 &
-    if [ -f /customer/lib/libpadsp.so ]; then
-        export LD_PRELOAD=/customer/lib/libpadsp.so
-    fi
+	printf "audioserver reactivado"
+	/mnt/SDCARD/Koriki/bin/audioserver
+	if [ -f /customer/lib/libpadsp.so ]; then
+		export LD_PRELOAD=/customer/lib/libpadsp.so
+	fi
 fi
